% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WESTGO.R
\name{calculateSMatrix}
\alias{calculateSMatrix}
\title{Calculate the similarity matrix}
\usage{
calculateSMatrix(gt, sampleNames = sampleNames, phased = T,
  minVariants = 5, scaleBySampleAF = F, outputDir = ".",
  saveResult = NA, varcov = T)
}
\arguments{
\item{genotypes}{data object containing the phased or unphased genotypes by samples}

\item{cex}{character expansion for the text}

\item{mar}{margin paramaters; vector of length 4 (see \code{\link[graphics]{par}})}
}
\value{
Object containing similarity matrix and other results
}
\description{
Calculate the similarity matrix
}
\examples{
data(toyGenotypes)
calculateSMatrix(toyGenotypes)

}

