% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stego.R
\name{plotStegoHeatmap}
\alias{plotStegoHeatmap}
\title{Plot the distribution of similarity statistics}
\usage{
plotStegoHeatmap(resObj, plotname = "", ...)
}
\arguments{
\item{plotname}{title of the plot}

\item{genotypes}{data object containing the phased or unphased genotypes by samples}
}
\value{
Object containing similarity matrix and other results
}
\description{
Plot the distribution of similarity statistics
}
\examples{
data(toyGenotypes)

run_stegotoyGenotypes)
labels <- c(rep("Group A",100), rep("Group B",100))
run_stegotoyGenotypes, groups="each.separately", labels=labels)

labels <- paste("Group",c(LETTERS[rep(1:4,25)],LETTERS[rep(5:8,25)]))
super <- c(rep("Super A",100), rep("Super B",100))
res <- run_stegotoyGenotypes, groups="pairwise.within.superpop", labels=labels, super=super)

}

