% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stego.R
\name{pruneGenotypes}
\alias{pruneGenotypes}
\title{Prune genotypes based on blocks of a specified length}
\usage{
pruneGenotypes(gt, blocksize = 1)
}
\arguments{
\item{gt}{data object containing the genotypes by samples}
}
\value{
Object containing similarity matrix and other results
}
\description{
Prune genotypes based on blocks of a specified length
}
\examples{
data(toyGenotypes)
pruneGenotypes(toyGenotypes)

}

