\name{stego-package}
\alias{stego-package}
\alias{stego}
\docType{package}
\title{
\packageTitle{stego}
}
\description{
\packageDescription{stego}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{stego}
\packageIndices{stego}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{stego}

Maintainer: \packageMaintainer{stego}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
