% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{summary.stego}
\alias{summary.stego}
\title{summary.stego}
\usage{
\method{summary}{stego}(object, ...)
}
\arguments{
\item{object}{an object of class "stego"}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Summary description of stego S3 object
}
\description{
summarizes the results of a stego analysis
}
\keyword{keywords}

