% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{summary.westgo}
\alias{summary.westgo}
\title{summary.westgo}
\usage{
\method{summary}{westgo}(object, ...)
}
\arguments{
\item{object}{an object of class "westgo"}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Summary description of westgo S3 object
}
\description{
summarizes the results of a WESTGO analysis
}
\keyword{keywords}

