% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stego.R
\name{plotFromGSM}
\alias{plotFromGSM}
\title{Plot the distribution of similarity statistics}
\usage{
plotFromGSM(resObj, plotname = "", alphaCutoff = 0.01)
}
\arguments{
\item{genotypes}{data object containing the phased or unphased genotypes by samples}

\item{cex}{character expansion for the text}

\item{mar}{margin paramaters; vector of length 4 (see \code{\link[graphics]{par}})}
}
\value{
Object containing similarity matrix and other results
}
\description{
Plot the distribution of similarity statistics
}
\examples{
data(toyGenotypes)

stego(toyGenotypes)
labels <- c(rep("Group A",100), rep("Group B",100))
stego(toyGenotypes, groups="each.separately", labels=labels)

labels <- paste("Group",c(LETTERS[rep(1:4,25)],LETTERS[rep(5:8,25)]))
super <- c(rep("Super A",100), rep("Super B",100))
res <- stego(toyGenotypes, groups="pairwise.within.superpop", labels=labels, super=super)

}

